Split out Question class to its own file.
